Added Async Methods
Reworked the necessary methods to use async and await. It works really well but i am unsure of the syntax. Also decided to implement further keywords like return and wait as functions, will add support for not using parenthesises next. if, elif, else, while and for are still operators in the parser.